reflow/local: don't retry image fetches indefinitely
Summary:
A recent change to Reflow's Docker utils introduced image pull
retrying. However, it's too aggressive: it only gives up if we get a
daemon error. This change limits the number of attempts to 5.

Additionally we fail fast if the image manifest or repository are not
found.

Fixes https://github.com/grailbio/reflow/issues/60

Reviewers: joshnewman, sbagaria

Reviewed By: joshnewman, sbagaria

Differential Revision: https://phabricator.grailbio.com/D16845

fbshipit-source-id: 0421554