Allow s3file to continue reading a file as long as its hash matches.
Summary:
Before, if a file was overwritten while someone was reading it, the reader would
get a "file version changed" error. This change relaxes the check so that the
reader can continue reading as long as the file's SHA256 hash remains the same.

This situation happens in GQL, for example. Ideally, the writer should be able
to do PutObject only if the file doesn't exist already, but such mode of
operation doesn't exist in S3 yet.

Reviewers: marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D17237

fbshipit-source-id: 8a091f5