reflow: tools for partial file resolution
Summary:
This change adds two new optional fields to reflow.File: Source
contains a source URL of the contents of the file, ETag contains an
ETag for that source. A Source should not be specified without an
ETag.

We also add a new method to Executor: Resolve elaborates a URL into a
reflow.Fileset with that contains "unloaded" files: they have a
Source, Size, and ETag, but not sha256 -- i.e., they don't have an
identity within Reflow itself.

Together, these provide tools to allow for (1) pipelining of loading
and execution; (2) passing URLs directly to execs where possible,
bypassing staging; and (3) cache assertions for sourced content.

The first is required to support a new scheduler that no longer
requires data to be staged on a "primary" alloc. The second can
improve performance of programs that can read directly from S3. The
third is to allow Reflow's caching to behave correctly with respect
to source input data that changes, e.g., directories or files on S3.

Reviewers: smahadevan, sbagaria, pgopal

Reviewed By: smahadevan, pgopal

Differential Revision: https://phabricator.grailbio.com/D18605

fbshipit-source-id: 1093e50