reflow/sched: scalable scheduler
Summary:
Package sched implements a new scheduler for Reflow tasks. Instead of
relying on a single "primary" alloc and work-stealing, it instead
schedules directly on to a set of allocs, creating them as needed
from an underlying cluster.

Combined with delayed file loading, this change will allow workloads
that violate the current set of data-locality assumptions embedded in
the work-stealing scheme. This scheduler should be able to support
large sets of tasks that are unrelated in terms of data locality.

The initial implementation is intentionally simple: tasks are matched
to allocs based on a simple bin-packing scheme, designed to try to
maximize the chance of freeing up allocs as offered load decreases.
Also unlike the current Reflow scheduler, the new scheduler makes use
of a repository to exchange input and output data: over time, we can
also try to optimize task assignment to mimize data movement.

Reviewers: smahadevan, sbagaria, pgopal

Reviewed By: sbagaria, pgopal

Differential Revision: https://phabricator.grailbio.com/D18910

fbshipit-source-id: 7aa854e