github.com/base/bitset: move bit iterator into its own package
Summary:
This moves the bitset scanner recently added to bio-spamtools to
a new base/bitset package.

(Planning to move CircBitmap under bio/ in an upcoming diff.)

Test Plan:
Unit test and benchmark included.  Benchmark code includes an
example of a much-higher-performance manually inlined loop
which can be employed when it really counts.

Reviewers: ysaito, ayip

Reviewed By: ysaito

Subscribers: cchang

Differential Revision: https://phabricator.grailbio.com/D19104

fbshipit-source-id: db164c1