base/s3file: determine bucket region through s3manager
Summary:
Previously we used the GetBucketRegion API call to determine a
bucket's region. However, this might require API permissions to use.
The SDK has an alternative way to determine a bucket's region, which
does not require permissions. (In fact, s3manager.GetBucketRegion
uses anonymous credentials for its API calls.) It appears that this
is done by inspecting the underlying HTTP redirects, using path-style
S3 requests.

The upshot is that this method always appears to work, regardless of
underlying permissions.

Reviewers: ysaito, joshnewman

Reviewed By: ysaito, joshnewman

Differential Revision: https://phabricator.grailbio.com/D19807

fbshipit-source-id: 10dd1a8