base/file: semantic errors
Summary:
Try to return semantic errors from file ops where possible. In
particular, we require that file.Open and file.Stat return an
errors.NotExist when a file is missing. We also change the S3 Open
implementation to check for existence on Open so that callers can
handle errors early.

This behavior is consistent with the os package, which returns
errors that can be checked with os.IsNotExist.

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D22424

fbshipit-source-id: 63a0171