base/simd: add non-amd64 fallback functions
Summary:
This adds non-amd64 implementations of all the simd API functions.

(The initial intent was to support a biogo PR dependent on one of these
functions, but it turns out that the function (reverse-without-complement)
is too unimportant to justify adding a dependency.  Still worthwhile to get
this package into shape, though.)

Test Plan: Existing tests pass with the build tags flipped to use the _generic.go code.

Reviewers: ysaito, sbagaria

Reviewed By: ysaito

Subscribers: cchang

Differential Revision: https://phabricator.grailbio.com/D23630

fbshipit-source-id: 6345668