cmdutil: do not link grail v23 factory
Summary:
Linking the grail v23 runtime factory registers flags and enables
flag parsing for v23 flags, which may be undesired in situations
when the code is being used as a shared library.

There is no workaround for such shared libraries because `flag.Parse`
can not be redefined, and it will be called if the v23 runtime
factory enabled flags.

`github.com/grailbio/base/vcontext` should only be linked in main packages.

Test Plan: Unit tests

Reviewers: ysaito

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D24793

fbshipit-source-id: e4b58b7