base/cmdutil: default ticket paths for TicketFlags
Summary:
This helps users of TicketFlags not have to put their own extra logic
for defaults and helps communicate to the users what the default values
are, and an intuitive way to clear these values to nil.

Test Plan:
Ran the internal grail-file command and was able to clear out any
default ticket paths by using an empty value for the command line flag.

Reviewers: ysaito

Reviewed By: ysaito

Subscribers: mskinner, jtoung

Differential Revision: https://phabricator.grailbio.com/D24932

fbshipit-source-id: b1d9c68