base/compress: change NewReaderPath to return io.ReadCloser
Summary:
I recently noticed that many uncompressing readers return io.ReadCloser, with a
caveat that a failure to call Close would miss file corruption and/or cause
memory leaks.

This diff changes the NewReaderPath to return io.ReadCloser.

There are many existing users of grailfile.UncompressingReader() that are
vulnerable to this problem, but I'll worry about them separately.

Reviewers: marius, xyang

Reviewed By: marius, xyang

Differential Revision: https://phabricator.grailbio.com/D24967

fbshipit-source-id: f7f823c