bio/spamtools: ordered pileup
Summary:
This adds a pileup mode which reports all base-quals and fragment
lengths in order (so one can determine which qual goes with which
fragment length).  The initial use case is on bio-collapse's output,
so stitching logic has been removed.

Test Plan:
Results on the existing test data appear consistent with the previous
pileup mode.  An end-to-end test will be added which verifies this.

Reviewers: ysaito, ayip

Reviewed By: ayip

Subscribers: rmauntz, jxiang, dburkhardt, schen, avalouev, cchang

Maniphest Tasks: T13303

Differential Revision: https://phabricator.grailbio.com/D24506

fbshipit-source-id: ebd2ab4