bio-spamtools: add 5'-distance, 3'-distance, strand columns to ordered output
Summary:
This extends the ordered pileup to be able to report a few more
statistics.  These were previously reported for ALT-supporting reads
by C-spamtool's "errlog" mode, but Ruth also needs REF.

(Note that read-pair stitching is NOT performed yet; if it's needed,
an option to perform that will be added in the next diff.)

Test Plan:
Existing tests pass.  Ran this on a test PAM and verified that output
was as expected.  Added a unit test verifying concordance with a
pair of hand-inspected output files.

Reviewers: ayip, ysaito

Reviewed By: ayip

Subscribers: rmauntz, cchang

Tags: #short_variants_development

Differential Revision: https://phabricator.grailbio.com/D27837

fbshipit-source-id: fb0d33f