s3/client: fix gratuitous cancellation during ticket retrieval
Summary:
The problem is that s3/client stashes a context object passed by a random caller
X, and later uses that context to issue an RPC to the Vanadium server.  If
caller X cancels the context, then all subsequent RPCs to vanadium will fail
immediately. In case of classifier-evaluation, X is file.RemoveAll, which
creates a context via errgroup.WithContext. That context is automatically
cancelled on error.

Reviewers: joshnewman, marius, jcharumilind, pgopal, sbagaria, O10 LIMS Server, rhuang

Reviewed By: marius, sbagaria, O10 LIMS Server, rhuang

Subscribers: ashanmugam

Differential Revision: https://phabricator.grailbio.com/D28798

fbshipit-source-id: d82cefd