file: add a flag to retry on S3 KeyNotFound error
Summary:
S3 may report KeyNotFound (or VersionNotFound) error when someone tried to GET the same file in a recent past.
This property makes it rather difficult for general applications to use S3, since they can't tell if anyone in the universe has ever tried to access the same file in the past.

To work around this problem, this change adds a flag to file.Open and file.Stat to retry automatically when S3 reports KeyNotFound. It also plumbs this flag to fragment & pam readers.

Reviewers: marius, joshnewman, ayip

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D30947

fbshipit-source-id: eeff832