base/simd: BitFromEveryByte
Summary:
This adds simd.BitFromEveryByte, which grabs the nth bit from every
byte and packs the result into a bitarray one-eighth the size.

The implementation is based on the _mm_movemask_epi8() SIMD
instruction, which is not currently exposed by the Go standard
libraries.  It has more than 50 times the throughput of a naive
algorithm, and 7 times that of the most sophisticated non-SIMD bit-hack
I'm aware of.

Ref T20646

Test Plan: Unit test and benchmark included.

Reviewers: ayip, ysaito, marius

Reviewed By: ysaito

Subscribers: cchang

Maniphest Tasks: T20646

Differential Revision: https://phabricator.grailbio.com/D30973

fbshipit-source-id: 73f491e