ticket-server: Allow ticket-server access to instance that are in a private subnet
Summary:
Relax the public IP validation for instances that are behind a NAT gateway and hence do not have a public IP.
Currently we check if the public IP is same as the remote IP they come from.
The fix will skip this check if the instance does not have a public IP.

Test Plan: Unit test passes. Will follow up with a test against a staging server.

Reviewers: aeiser, ysaito, bbentson, marius

Reviewed By: aeiser

Maniphest Tasks: T23459

Differential Revision: https://phabricator.grailbio.com/D32675

fbshipit-source-id: 1592cbb