regionfrags: add flexible way to write chopping distributions, add tsv driven fixed distribution
Summary:
This change refactors how a chop distribution is constructed to make
it easier to add a new distribution.

This change replaces TrainMeanChopLength and PredictMeanChopLength with TrainChopDistribution and PredictChopDistribution.

These two parameters are strings that look like this:

"exponential&10" (where 10 is the mean length)
"fixed&distribution.tsv" (where distribution.tsv is the path to a file describing a distribution)

Test Plan: Added unit test.

Reviewers: marius, rcalef

Reviewed By: rcalef

Subscribers: nshojaee, rcalef

Maniphest Tasks: T22302

Differential Revision: https://phabricator.grailbio.com/D33452

fbshipit-source-id: 64f4fa8