s3file: make sure to set server side encryption flag in PutObject.
Summary: Previously, even if server side encryption was specificed in Options, it was not being set in PutObject request.  This only presented itself when writing/uploading an empty file, becuase the uploader handles that with a special case.

Test Plan: Tested manually by trying to use grailfile to write an empty file to s3 before and after the change.

Reviewers: ysaito, marius

Reviewed By: ysaito

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D35062

fbshipit-source-id: 1366bdd