bigmachine: detect and propagate OOM errors
Summary:
This change adds OOM detection to bigmachine. It works by adding a
new facility to System that allows the supervising node to read the
system's kernel message buffer. It then looks for messages indicating
that the bigmachine process was killed by the kernel OOM killer, and
errors the machine accordingly.

We also adjust the OOM score of the bigmachine process so that the
kernel preferentially kills it (and not, say, our monitoring of
kmsg). We also add a small shutdown delay to give the supervisor some
time to detect the OOM.

With this change, running into an OOM condition with Bigslice yields the
following error, which is both unambiguous and much more user friendly
than the current heuristics used.

	2019/09/05 17:05:23 machine https://ec2-34-222-254-154.us-west-2.compute.amazonaws.com:2000/ (ok) stopped with error bigmachine process 874 was killed by the kernel OOM killer

As a follow-up change, since we now return a semantic error for
process OOMs, we can change Bigslice to dynamically adjust its load
factor when these are encountered, so that future tasks have a better
chance of survival.

Test Plan: Run new test program, `bigoom`.

Reviewers: jcharumilind, joshnewman, pgopal

Reviewed By: jcharumilind, joshnewman

Subscribers: smahadevan, jschellenberger

Differential Revision: https://phabricator.grailbio.com/D35921

fbshipit-source-id: 512c848