file/s3file: Presign method
Summary:
The new method can be useful to transfer files to clients across RPCs
along with limited time credentials, instead of transferring blobs as
the RPC payload.

It can also be used in reflow execs to provide a signed URL to the exec
instead of a local copy of the file from the reflow repository.

Test Plan: Ran the new unit test manually with an S3 bucket.

Reviewers: smahadevan, ysaito, marius

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D36178

fbshipit-source-id: bb16813