base/fatbin: implement a simple fat binary format
Summary:
Package flatbin implements a single fat binary format, providing an
API to create and read such binaries. Fatbin binaries are not
traditional fat binaries, but instead rely on having a base binary
(ELF or Mach-O), to which a zip archive is appended with the
remainder of the supported architectures.

Fatbin will be used to support using bigmachine-based systems to be
run from macOS, and potentially others. We can also use this in
Reflow, in places of its current mechanism (which is limited to
darwin+linux).

Reviewers: pgopal, smahadevan, jcharumilind, ysaito

Reviewed By: smahadevan, ysaito

Differential Revision: https://phabricator.grailbio.com/D36271

fbshipit-source-id: 5345e59