grail-access: refresh credentials as default behavior
Summary:
Current behavior is to merely dump the current principal and print the
expiry of the default blessing if the expiry time is more than 30
minutes from now.

This behavior will now be behind a new flag `--dump` and the expiry time
check will be increased to 7 days to keep the current behavior for long
lived credentials (e.g. EC2 creds).

Test Plan:
Manually ran the command with and without the flag, and with and without
the credentials directory present.

Reviewers: aeiser, marius

Reviewed By: aeiser

Subscribers: pknudsgaard, marius, smahadevan

Differential Revision: https://phabricator.grailbio.com/D36598

fbshipit-source-id: 433746a