fatbin: make the format more robust
Summary:
Fatbin currently relies on the sniffers to be able to determine the
offset of the fatbin zip. This is done by inspecting the image format
directly. This approach has not proven to be robust. While it seems
to work on common configurations inside of GRAIL, we have reports
(e.g., issue #8) of the approach failing in other configurations.

This change changes the format to include a magic value to identify
fatbin binaries, and a footer that stores the fatbin offset directly.
This should prove a much more robust approach as we do not rely on
accurately reading the underlying image format with all of of its
implementation-specific variations.

Closes #9.

Reviewers: pgopal, smahadevan, jcharumilind, ysaito

Reviewed By: smahadevan

Differential Revision: https://phabricator.grailbio.com/D37658

fbshipit-source-id: 22487a9