ticket-server: switch to simpler group.get api call to fetch cross account
Summary:
T7463 switched the ticket-server to use HasMember API call that does lookups for neseted inheritence.

T15539 switched the ticket-server to use cross account lookups, so a user with a @grailbio.com email could be in a group from summitstudies.co.uk and still be validated as a member.

T29527 noticed that "sometimes" due to a Google API delays - the existing cross account lookup is slow.

This diff reverts the cross account lookup to use the older group.get(member) API call which should avoide the API issues, and is a simplification over the old code.

Test Plan: Verify in SBX staging, and create a test cross group lookup.

Reviewers: #shared_infrastructure, trooney

Reviewed By: trooney

Subscribers: sbagaria, joshnewman, rhuang, smahadevan, mfoster

Maniphest Tasks: T29527

Differential Revision: https://phabricator.grailbio.com/D39771

fbshipit-source-id: 335e2a8