grail-access: blackbox tests
Summary:
I'd like to make some changes to grail-access. I've wondered for a long time why
it deletes an existing Vanadium principal before it starts an authentication
flow instead of just updating the blessing. I partly think it's inelegant, but
I've also seen multiple problems in #eng-help that were caused or exacerbated
by running grail-access, canceling it, and not realizing that destroys the
credentials.

After a bit of code inspection recently, I think grail-access has quite a bit
of historical cruft in its code (handling of the principal and its blessings is
disorganized and redundant), and I think the patches to work around agent errors
could be fixed in a more principled way.

However, since we use it a lot and I'm learning as I go along, my first step is
to introduce some tests. These all invoke grail-access as if from a shell and
check its behavior and output to ensure user-visible behavior is unchanged.

The connection of this to my actual work is rather distant: I am hoping this
fixes a grail-fuse issue which will make it easier to do non-Go analysis work
and otherwise streamline some dev environment tasks and then I'll be so much
more productive (ha). I think it's worth giving grail-access some attention,
since it is almost in its initial prototype state from three years ago, but it
might just take some time, and I understand if it takes time to review.

Test Plan:
Tests the EC2 blesser. I don't know how to reasonably fake the Google oauth2
server, but that's also the more reasonable flow to test manually, so we may
need to just rely on that.

Tests pass with both Bazel and `go test`.

Reviewers: aeiser

Reviewed By: aeiser

Subscribers: smahadevan, krestivo, afields, sbagaria

Differential Revision: https://phabricator.grailbio.com/D44965

fbshipit-source-id: e2ff93d