bio/demuxbcl: allow 256+ samples
Summary:
The first version of demuxbcl was limited to 255 samples, since it was
convenient to use the "bytes" package and gunsafe.ExtendBytes() for a
few operations.  However, since we now know we need to demux larger
numbers of samples at a time, this revision changes the sampleId type
to uint16, which raises the limit to 65535 samples.

bytes.IndexByte() was actually moderately important for performance,
so I've added the analogous simd.IndexU16() function.

Test Plan:
Existing tests pass, and unit test and benchmark have been added for
simd.IndexU16().  Reran bio-demux-bcl on the dataset mentioned in
T34299, and verified results were unchanged.

Reviewers: ayip, ghorrell

Reviewed By: ayip

Subscribers: smahadevan, escott, sramani, nshojaee, cchang

Maniphest Tasks: T39842

Differential Revision: https://phabricator.grailbio.com/D48010

fbshipit-source-id: d47d8ab