Changed vlog to common logger
Summary:
Changed vlog to common logger
Where ever the context was not available in "main" like function, context.Background() context was used. In the NoCtx variations of the logger, the context.Background() is used (suggested by @treaster)

Test Plan: This should be printing the same message as original message. No change in functionality is expected.

Reviewers: noah, O56 Ensemble backend, fdegiuli

Reviewed By: O56 Ensemble backend, fdegiuli

Subscribers: noah, treaster, smahadevan, fdegiuli

Differential Revision: https://phabricator.grailbio.com/D51822

fbshipit-source-id: 776f0ce