grail-ticket,ticket-server: fix empty string glob issue
Summary: Previously, registering the empty string "" endpoint was causing glob to fail. So, we will change the endpoint to be "list" and append "list" to the path.

Test Plan:
Deployed locally and to staging tickets-testing ticket-server and verified flow with:
```
grail-ticket -list tickets-testing
```
Also verified that the glob functionality still works:
```
namespace glob tickets-testing/...
```

Reviewers: O9 ticket-server, bbentson

Reviewed By: O9 ticket-server, bbentson

Subscribers: aeiser, bbentson, smahadevan

Differential Revision: https://phabricator.grailbio.com/D53226

fbshipit-source-id: 85c3615