log: caller field should only be file name and line
Summary:
Change the automatically logged `caller` field from full file path to only file name.

The impetus for this change is that the caller root will be different depending on the system. For example, the tests will pass with `bazel test`, but fail with `go test`. We could possibly configure the full file path to be an option in the future.

Github Issue: https://github.com/grailbio/base/issues/28

Test Plan: Update unit tests.

Reviewers: fdegiuli, msabherwal

Reviewed By: fdegiuli

Subscribers: jcharumilind, smahadevan

Differential Revision: https://phabricator.grailbio.com/D53636

fbshipit-source-id: 0a74d6d