log: add logger interface
Summary:
This diff refactors the logging system to support an interface. This gives more flexibility to users, specifically the ability to mock the logger.

With this refactor, we now support two logging flows. The first is the current flow with global, structless functions. The second is a flow that requires instantiation of the logger via `NewLogger()`.

Test Plan: Update unit tests and include tests for both flows.

Reviewers: msabherwal, fdegiuli

Reviewed By: msabherwal, fdegiuli

Subscribers: aeiser, jcharumilind, smahadevan

Differential Revision: https://phabricator.grailbio.com/D53655

fbshipit-source-id: 0a27a40