log: add logging config
Summary:
Currently, due to output being piped to both stderr and stdout, the logger package duplicates logs when run within a Kubernetes pod. The logger package does not duplicate logs when run within an EC2 instance because only stderr is captured.

This diff will fix the issue of logs being duplicated when run within a Kubernetes pod by adding the ability to configure the output path.

This diff also lays the foundation of customizing the logger. In a separate diff, I will add the ability to set the log level, which currently defaults to `Debug`.

Test Plan: Update unit tests.

Reviewers: fdegiuli

Reviewed By: fdegiuli

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D55332

fbshipit-source-id: c75cb96