recordio: Add SkipHeader and KeyTrailer to WriterOpts
Summary:
This change adds a `SkipHeader` option that allows for users to create a writer that skips writing a header and begins in the `wStateWritingBody` state. This can be used to continue writing from where a previous writer may have left off.
This change also adds a `KeyTrailer` option that is needed when the recordio file contains a trailer. Previously the only way to set this was by explicitly adding this as a header.

Test Plan: Unit tests have been added for the normal write then read functionality as well as testing functioning with transformers.

Reviewers: kshashidhar, smahadevan

Reviewed By: kshashidhar, smahadevan

Subscribers: pbohannon

Maniphest Tasks: T47010

Differential Revision: https://phabricator.grailbio.com/D54947

fbshipit-source-id: be5a000