bazel/generated: rule for managing checked-in generated code
Summary:
Use new rule to manage generation of fragments code. I think this will be helpful for maintainability (for non-github.com code).

If this works well, future changes can expand it to the rest of the generated fragments code, protobuf stubs (so we don't need to handle those separately in Arcanist), SQL schemas, VDL, cmdline doc.go, etc.

Test Plan: "Accidentally" modified a generated file, got a prompt to regenerate it, and then fixed it.

Reviewers: jcharumilind

Reviewed By: jcharumilind

Subscribers: noah, smahadevan, sbagaria

Differential Revision: https://phabricator.grailbio.com/D55482

fbshipit-source-id: 7e1908a