common/log: set default log level to debug
Summary:
Currently, the default log level is `Info`. This is because the Zap library sets the default to `Info` and an empty level corresponds to `Info`. Any instantiation of `Config` will have `Level` as `Info` instead of `Debug`.

This is a messy problem and I think this is the best solution. If you can think of a better one, please let me know!

Test Plan: .

Reviewers: fdegiuli, jkimmel

Reviewed By: jkimmel

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D56367

fbshipit-source-id: 3ba967e