common/middleware : Introduces consensus-driven middleware folders and Introduces Debug logs for v23 calls
Summary:
Middleware approach allows this invoker to wrap existing invokers.
full V23 input args and results/errors are reported to logs.

This approach allows us to turn on/off highly verbose logging, including of sensitive information,
by toggling to Debug log-level.  If there comes a day when we still want the option to run in Debug
in production but exclude these logs we can either add a feature-flag / env var check, or remove
this component entirely.

This change also shows how the logger is added to the notifier service.
adding this logger to other services can be done in separate diffs.

[CPP-1695]

Test Plan: Manually tested by deploying notifier to dev and observing the logs

Reviewers: psteed, fdegiuli, O50 galleri/communications, lkong, akshay.subram, noah, nyap, O63 logger

Reviewed By: fdegiuli, O50 galleri/communications, lkong, O63 logger

Subscribers: smahadevan, O59 lockbox, noah, ccoakley

Differential Revision: https://phabricator.grailbio.com/D56577

fbshipit-source-id: 7bc46a9