base/iofmt: add LineWriter to call Write with whole lines
Summary:
`PrefixWriter` makes separate calls to the underlying writer's `Write` for the prefix and line contents.  This makes it more likely for lines to become interleaved when there are multiple writers to the same underlying writer, e.g. `os.Stdout`.

This adds a `LineWriter` that uses a (unbounded) buffer to only call `Write` with full lines.  This can be composed with a `PrefixWriter` to reduce interleaving.

We can still get interleaving, as the call to `Write` is not atomic, however this is a practical improvement (without coordination like locking).

Test Plan:
Run this test program and see that there is less (zero) interleaving.

```
package main

import (
	"fmt"
	"os"
	"sync"

	"github.com/grailbio/base/iofmt"
)

func main() {
	const (
		Nwriters = 20
		Nlines   = 100
	)
	var wg sync.WaitGroup
	wg.Add(Nwriters)
	for i := 0; i < Nwriters; i++ {
		i := i
		go func() {
			defer wg.Done()
			pfx := fmt.Sprintf("writer-%02d: ", i)
			w := iofmt.PrefixWriter(iofmt.LineWriter(os.Stdout), pfx)
			// w := iofmt.PrefixWriter(os.Stdout, pfx) // Interleaves!
			for j := 0; j < Nlines; j++ {
				line := fmt.Sprintf("line %03d\n", j)
				if _, err := w.Write([]byte(line)); err != nil {
					fmt.Fprintf(os.Stderr, "writer-%02d error: %v\n", i, err)
					return
				}
			}
		}()
	}
	wg.Wait()
}
```

Reviewers: joshnewman

Reviewed By: joshnewman

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D57169

fbshipit-source-id: 7091742