[SYSINFRA-367,SYSINFRA-185] base/limiter: Prevent panics if there are no ids to claim
Summary:
Remove a panic when there are no ids to claim.
This can happen with a liberal limiter, as one of the goroutines can claim the ID,
and the other ones won't find anything to claim (since the limiter allowed them to make calls).

@jirgon found this in the nightly tests:
```
2021/06/05 14:04:34 ec2cluster: [i-0600e7eeccbe97b54]: terminating: pending -> shutting-down
panic: Do(i-0600e7eeccbe97b54): this is a bug - no ids for batch call
```

Test Plan:
Unit tests.

With just the changes to `batch_test.go`, the unit test fails:
```
> go test -run SometimesDedup
panic: Do(test): this is a bug - no ids for batch call
```

With the changes to `batch.go`, the unit tests pass:
```
> go test -race -test.count 10000 -run SometimesDedup
PASS
ok  	github.com/grailbio/base/limiter	7.475s

```

Reviewers: ghorrell

Reviewed By: ghorrell

Subscribers: jirgon

Differential Revision: https://phabricator.grailbio.com/D60074

fbshipit-source-id: d0d9e5f