Revert "[SYSINFRA-574] pipeline/rds: fixes to increase db-reader role ttl"
Summary:
This reverts commit 79d31b7ce5eaca3291ee9cc9cd993b724caab000.

As noted in [this thread](https://grailbio.slack.com/archives/CAA39NK7Y/p1628541490082300), when D62995 was landed with `arc land`, it somehow included a revert of a lot of previous commits. It seems like some problem occurred during the merge phase of `arc land`, but it didn't display any error messages.

I’ve added the authors of all affected diffs as reviewers to the revert diff.

{F1455817}

Test Plan: N/A

Reviewers: O9 ticket-server, O58 Shipment Service owner, O60 Accounts Service Backend, O61 Portal, apaniagua, amitchell, cdonohoe, smahadevan, chsu, sbagaria, asoni, jirgon, adharwadkar

Reviewed By: O9 ticket-server, O58 Shipment Service owner, O60 Accounts Service Backend, O61 Portal, apaniagua, amitchell, cdonohoe, smahadevan, chsu, sbagaria, asoni, jirgon, adharwadkar

Subscribers: asoni, smahadevan, dnicolaou

Differential Revision: https://phabricator.grailbio.com/D63004

fbshipit-source-id: 347d7f7