ticket-server: Add type-safe helpers
Summary:
This adds helpers for all of the ticket types exposed by
`github.com/grailbio/base/security/ticket` plus byte slices and strings
(which are normally accessed via `GenericTicket`). In addition to taking
care of casting the `ticket.Ticket` interface to the appropriate type
(and handling errors when this fails), the API is designed to work with
a pattern observed in the ticket paths for microservices deployed in
multiple Kubernetes clusters.

It seems that a common pattern for ticket paths is
`tickets/aws-grail-eng-$env/$service/$key` where the same $service will
access one or more tickets which share a prefix that varies by
environment / cluster name. As it stands, the entire path must be
specified for each ticket as an environment variable in the Helm charts
(and is usually passed in as a flag). Instead, it should be possible to
pass in just the shared prefix `tickets/aws-grail-eng-$env/service` and
allow the service (or this library) to construct the rest of the key.

Both approaches are handled by the helpers by using a variadic argument for
the ticket path. This pattern seems to be broken by the new `preprod`
tickets, however, so perhaps this will not be used in practice.

The new API looks like

```
myTicketString, err := ticket.Client.GetString(v23context, "path/to/ticket")
return myTicketString, err
```

instead of

```
tsClient := ticket.TicketServiceClient("path/to/ticket")
tick, err := tsClient.Get(ctx)
if err != nil {
  return "", err
}
generic, ok := tick.(ticket.TicketGenericTicket)
if !ok {
  return "", fmt.Errorf("uh oh")
}
return string(generic.Value.Data), nil
```

Test Plan:
  - Unit tests cover type casting and some of the error handling.
  - Integrated into the `accesshope` service and was able to get
    tickets. This will come in a follow up diff if this is accepted.

Reviewers: bbentson, zpallin, noah

Reviewed By: noah

Differential Revision: https://phabricator.grailbio.com/D62369

fbshipit-source-id: 3881f3c