fix error message for opening namespace and token files
Summary:
When running `grail-access` with either `-namespace=unknownfile` or `-token=unknownfile` the error provided referenced "opening ca.crt". This could lead to confusion that the flags are not being correctly parsed.

This change fixes the error messages.

Test Plan: Ran `grail-access` with the above commands and confirmed that the appropriate error message is shown

Reviewers: O68 security, noah

Reviewed By: O68 security, noah

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D63956

fbshipit-source-id: a6a22d8