[BXDS-1071] grail-fuse: kill DirPlusStream
Summary:
We have our own fork of `go-fuse` which adds a `DirPlusStream` interface used to make LOOKUP FUSE operations during READDIRPLUS more performant.  (Without some mitigation, we make O(n) calls to S3 when listing a directory: one for each LOOKUP for which there is one for every directory entry.)

We started the process to upstream it, but the `go-fuse` maintainer pointed out some issues, and we eventually abandoned the change/effort.  See more context in the [[ https://review.gerrithub.io/c/hanwen/go-fuse/+/524431 | code review ]].

Kill our use of this fork/patch to avoid the noted problems, and allow the previously returned dirstream entry/inode to be reused to make LOOKUPs fast.

One of the uses of the patched version was our internal `grail-fuse` tool which has now been superseded by `biofs`.  Kill it too so that we do not have to upgrade it.

Test Plan: Run with debug logging.  Use `ls` and observe that LOOKUP reuses nodes and does not trigger API calls.

Reviewers: smahadevan, joshnewman

Reviewed By: joshnewman

Subscribers: smahadevan, dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D67249

fbshipit-source-id: ac84bf6