bio/pileup/indel: first working version
Summary:
This fills in the remaining pieces (shard-processing and TSV-output
logic) for the indel-supporting pileup.

(In the meantime, yet another bit of code that was suppressing PerRead
output for N-supporting reads has been removed.  This finally broke
the bio-pileup and bio-spamtools tests in the expected manner; those
test fixtures have been updated.)

[BXDS-942]

Test Plan:
Unit tests added for full pileup workflow.  Ran this on a 2.3 GB test
BAM with and without -cols=+dpalt, and with and without
-per-strand=true, and verified that results looked sane.

Also ran this on
s3://grail-results/611183/cfDNA1_10_B9_GPSL_25pM_24Plex/bam/dupmarked.merged.bam ;
this resulted in a cryptic panic due to the presence of irregular
CIGAR strings (e.g. there are reads where the CIGAR ends in an
insertion rather than a soft-clip).  If others are likely to try this,
we should either make bio-pileup crash with a more informative
error message, or just permit irregular CIGARs.

Reviewers: dnicolaou, ckuo

Reviewed By: dnicolaou, ckuo

Subscribers: sbagaria, escott, cchang, smahadevan

Differential Revision: https://phabricator.grailbio.com/D68533

fbshipit-source-id: 07d91e1