fsnodefuse: handle duplicate directory entry names without panic
Summary:
Handle duplicate directory entry names without panicking.  We attempt to reuse `go-fuse` inodes (mostly to cache during handling of READDIRPLUS) and were panicking because we would attempt to reuse a file inode as a directory inode.  We now make sure that the modes match before reusing the inode.

Note that this kills the panic but still leaves pretty weird behavior when we have duplicate names, i.e. showing multiple entries with the same name with no way of differentiating them.  That will be addressed in a separate revision.

[BXDS-1452]

Test Plan: Run added unit test.  Run `biofs` and see that listing a "directory" in S3 with a file and directory with the same name, e.g. `ls /mnt/biofs/s3/grail-joshnewman/tmp/dir/`, does not cause any panics.

Reviewers: joshnewman

Reviewed By: joshnewman

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D71377

fbshipit-source-id: e286f0f