Revert "Golang: upgrade to 1.18"
Summary:
This reverts commit cb6090e46d7d674b7b827b9de096aa9d77b8a45a.

From slack:

```
ghorrell:
Hey @sngoc - we’re seeing increased memory usage on some reflow integration tests [1] which we believe is related to this version bump (see the notes on GC in the version notes for 1.18). Would it be possible to revert the 1.18 version upgrade? These tests passing is a dependency for a release branch cut that we need to make tomorrow. (CC: @olga)

https://buildkite.com/grail/reflow-integration/builds/2371
```

Also:
```
Kai Wells
fwiw upgrading to 1.18 also broke the LIMS database connection because the version of MySQL is too old to support TLS 1.2: https://grailbio.slack.com/archives/CAA39NK7Y/p1647558600777319
```

Reviewers: O45 fragments, ghorrell, oignatova, joshnewman, sdunn, kwells, dborcherding

Reviewed By: O45 fragments, ghorrell, joshnewman

Subscribers: smahadevan, joshnewman

Differential Revision: https://phabricator.grailbio.com/D72202

fbshipit-source-id: 1ba1ab6