base/errors: map errors.NotSupported to syscall.ENOTSUP
Summary:
Map `errors.NotSupported` to `syscall.ENOTSUP` instead of `syscall.ENOSYS`.

The available error kinds in `errors` is less granular than those of `syscall`.  According to [[ https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/errno.h.html | POSIX ]]:
```
[ENOSYS]
    Functionality not supported.
[ENOTSUP]
    Not supported (may be the same value as [EOPNOTSUPP]).
```

My understanding is that in practice `ENOSYS` signals that an entire function(ality) is not supported whereas `ENOTSUP` signals that some configuration of that functionality is not supported, e.g.
- A file abstraction may not support renaming at all, so requests to rename return `ENOSYS`.
- A file abstraction supports renaming only if the file is backed by a local regular file, so requests to rename a non-local file return `ENOTSUP`.

`errors` doesn't provide this distinction, so prefer the more granular mapping.  This may end up wrongly suggesting that there exists some configuration of a functionality that may work, but I think the (unlikely?) outcome is that we try operations that we might have known not to try earlier.

Test Plan:
Run `errors` unit tests.
Run `fsnode` and `fsnodefuse` unit tests, as these are the main users of this mapping.

Reviewers: joshnewman

Reviewed By: joshnewman

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D72841

fbshipit-source-id: 3e75a09