file/s3file: parallel S3 lister
The command line tool writes the same output format as grail-file ls (to allow comparison, for testing).

The tool and its implementation should be considered experimental for now. I think it'd require more usage and testing to be confident the results match the straightforward serial version.

If we like this, we can make this non-experimental (perhaps with some API polish).

Test Plan:

Listed contents of some directories and buckets (with and without -l and -R) and compared results to the serial grail-file ls.

On an internal bucket with ~500k key prefixes at the top level, a shallow listing (of "directories") was ~9x faster (1m54s -> 13s), producing the same results.